#78 [refactor] : Chatbot 도메인을 DDD 구조로 리팩토링한다
