+++ src/main/java/kusitms/backend/chatbot/application/dto/request/MessageDto.java
@@ -1,4 +1,4 @@
-package kusitms.backend.chatbot.dto.request;
+package kusitms.backend.chatbot.application.dto.request;
 
 public record MessageDto(
         String role,
